Cleaning up model debug work. Preparing to merge into master
