+++ python/setup.py
@@ -50,14 +50,14 @@ extensions = [
               sources = ['underworld3/stokes.pyx',
                          'underworld3/functions.c'],
               **configure()),
-#    Extension('underworld3.poisson',
-#              sources = ['underworld3/poisson.pyx',
-#                         'underworld3/poisson_setup.c'],
-#              **configure()),
-    Extension('underworld3.pyjules',
-              sources = ['underworld3/pyjules.pyx',
-                         'underworld3/poisson_jules.c'],
+    Extension('underworld3.poisson',
+              sources = ['underworld3/poisson.pyx',
+                         'underworld3/poisson_setup.c'],
               **configure()),
+#   Extension('underworld3.pyjules',
+#             sources = ['underworld3/pyjules.pyx',
+#                        'underworld3/poisson_jules.c'],
+#             **configure()),
 ]
 
 setup(name = "underworld3",
